requirements.txt
requests
boto3
python-dotenv
beautifulsoup4
aiohttp>=3.9.0
aioboto3>=12.0.0
tenacity>=8.2.0
cdp-backend
fireo
gcsfs
google-cloud-firestore
tqdm>=4.66.0
groq
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
slowapi>=0.1.9
pydantic>=2.0.0
matplotlib>=3.7.0
pillow>=10.0.0
llama-parse
sentence-transformers
langchain-classic
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
chromadb
anthropic
numpy
scipy
langchain-anthropic
langgraph
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
